# Add shortcuts for missing data