--- a/pomegranate/distributions/JointProbabilityTable.pyx
+++ b/pomegranate/distributions/JointProbabilityTable.pyx
@@ -210,25 +210,22 @@ cdef class JointProbabilityTable(MultivariateDistribution):
 
 	cdef double _summarize(self, double* items, double* weights, int n,
 		int column_idx, int d) nogil:
-		cdef int i, j, idx, is_na
+		cdef int i, j, idx
 		cdef double count = 0
 		cdef double* counts = <double*> calloc(self.n, sizeof(double))
 
 		memset(counts, 0, self.n*sizeof(double))
 
 		for i in range(n):
-			idx, is_na = 0, 0
+			idx = 0
 			for j in range(self.m+1):
 				if isnan(items[self.m-i]):
-					is_na = 1
-
+					break
 				idx += self.idxs[i] * <int> items[self.m-i]
 
-			if is_na == 1:
-				continue
-
-			counts[idx] += weights[i]
-			count += weights[i]
+			else:
+				counts[idx] += weights[i]
+				count += weights[i]
 
 		with gil:
 			self.count += count