@jmschrei 2019-07-29
This looks good. I'll review it more thoroughly when I get back next week from the conference I'm at. Would it be possible for you to confirm that you're getting the same network before and after your changes by printing out `model.structure` and by calculating the probability of some data set? It can be the one that learned the network over. Thanks!