@jmschrei 2019-08-07
When I run the test script on my machine I don't see the same speedups that you see.

This is the result before incorporating your change to `discrete_score_node` (which is generally the most compute intense part of structure learning).

```
Without missing values:
      cpu_time  wall_time                                                                                                                                                                
mean  5.683641   8.551788
max   5.730023   8.625430
std   0.068112   0.064195

With missing values:
      cpu_time  wall_time                                                                                                                                                                
mean  5.138517   7.745014
max   5.150633   7.810688
std   0.016490   0.057186
```

When I incorporate the change:

``
Without missing values:
      cpu_time  wall_time                                                                                                                                                                
mean  5.744121   8.610118
max   5.888380   8.690713
std   0.137677   0.070891

With missing values:
      cpu_time  wall_time                                                                                                                                                                
mean  5.517163   8.523160
max   5.574747   8.530496
std   0.066018   0.012541
```

To me, the differences look like normal run-to-run differences. 

When I cranked up the percentage of missing values, I actually got a segfault in my code. It took me a while to figure it out, but the issue seems to have to do with the score calculation of a candidate parent set, which is P(D|M) + log(|D|) / 2 * |M|, where |D| is the number of examples in your data set and |M| is the number of parameters in the model. Pomegranate handles missing data when building Bayesian networks by ignoring examples that have missing values when focusing on the subset of features corresponding to a variable and it's candidate parents. This means that |D| can change from one candidate parent set to the next. In cases where the candidate parent set is large and the data full of missing values, you may get situations where there is only one example in the data where no features are missing. In this case, P(D|M) must be 0, but log(|D|) is also 0, meaning that complexity is no longer penalized. The segfault appears to be related to this, where massive candidate sets were being selected for some variables. The fix, derived from the MDL regularization itself, appears to be to discard candidate parent sets which result in only one example, much like we discard candidate parent sets that result in zero examples.



Regardless, after incorporating this simple fix, I timed both the original code and the code incorporating this change and removing ~99% of the values. The original code:

```
Without missing values:
      cpu_time  wall_time                                                                                                                                                                
mean  5.665113   8.461362
max   5.768124   8.482762
std   0.103726   0.021463

With missing values:
      cpu_time  wall_time
mean  0.171360   0.170214
max   0.172353   0.171646
std   0.000871   0.001241
```

The proposed changes:

```
Without missing values:
      cpu_time  wall_time                                                                                                                                                                
mean  5.484248   8.345044
max   5.521972   8.416701
std   0.032916   0.070389

With missing values:
      cpu_time  wall_time
mean  0.160196   0.160249
max   0.161104   0.162660
std   0.001142   0.002100
```

The difference in speed between the original code and the modifications appears minor, if not just run-to-run variation. Are you sure that these changes are causing a dramatic speedup on your end? To be fair, I didn't incorporate the changes to ConditionalProbabilityTable or JointProbabilityTable because I didn't expect them to make as much of a difference as this function.