@jmschrei 2019-08-07
I added in all of your changes and get the following timings when using ~99% sparsity:

```
Without missing values:
      cpu_time  wall_time                                                                                                                                                                
mean  5.520043   8.370361
max   5.609389   8.449367
std   0.123448   0.125639

With missing values:
      cpu_time  wall_time
mean  0.159223   0.160261
max   0.160032   0.167227
std   0.000870   0.006134
```

When I go back to ~10% sparsity I get the following, which is comparable to the timings above:

```
Without missing values:
      cpu_time  wall_time                                                                                                                                                                
mean  5.564584   8.379262
max   5.654466   8.494476
std   0.125364   0.137579

With missing values:
      cpu_time  wall_time                                                                                                                                                                
mean  5.649820   8.333919
max   5.746899   8.389962
std   0.109258   0.059623
```