@alexhenrie 2019-08-07
Or could your fix for data with too many missing values be affecting the performance? Could you put that fix on a branch where I can look at it?