@jmschrei 2019-08-07
Each time I would modify the code locally, build it, and then run the timings twice to ensure that there weren't any first-run issues. I'm running Python 3.7.3 GCC 7.3.0 in an Ubuntu vmware virtual machine on an i7-8700K 6 core (12 thread) CPU @ 3.70GHz. 

My fix involved simply adding 

```python
	if w_sum > 1:
		logp -= _log(w_sum) / 2 * m[d+1]
	else:
		logp = NEGINF
```

to the end of the `discrete_score_node` function, so I doubt that it was significantly impacting the speed of that function. It is possible that, by returning NEGINF rather than 0, that it is building a smaller network when the data has many missing values and so the time taken by `ConditionalProbabilityTable` is lessened. However, since my script segfaults with the original code, I haven't been able to time that.