@alexhenrie 2019-08-07
I don't see any improvement when running in a virtual machine either. I think you have to be running on bare metal to see the difference.